fix: move navbar to reusable components
